[docs] Add "See Also" blocks for `torch.t` and `torch.transpose`
**Summary**
PR 63177 added "see also" cross-references to `torch.t` and
`torch.transpose`. However, it did so in the form of simple links
preceded by the phrase "See also". A comment on the aforementioned PR
pointed out the RST has special syntax to render a "See Also" box that
is more noticeable. This commit converts the "see also" sentences for
`torch.t` and `torch.transpose` to boxes.

**Test Plan**
Continuous integration.

**Fixes**
This commit fixes #56267.

ghstack-source-id: 2b8eda6f774fc7b57c583b840bd4addf63cc6238
Pull Request resolved: https://github.com/pytorch/pytorch/pull/63461