[quant][fx] Don't assume bias is a keyword argument
`prepare_fx` currently assumes that bias is always a positional argument to
convolutions, and only a key argument to other functions. This happens to work
today due to a quirk in how `__torch_function__` is handled for python
functions but shouldn't be considered stable.

Instead, we should support `bias` for both positional and keyword forms.

ghstack-source-id: 9ca55ff420d4200ee2690dfe1687edcc6433d308
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61647