[TensorExpr] Add lowering for aten::max (reduction).
Is it a hack?
Yes and no. Yes, because I had to disable lowering for pointwise max.
No, because eventually we would need to add max reduction lowering
anyway. A proper implementation requires #65843 - that would allow us to
distinguish pointwise case from the reduction case and have separate
lowerings for them. In the long run we should land #65843 and then a
cleaned up version of this PR.

ghstack-source-id: b1e29a20a1410956c040a2612f0dbf3499581f7f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66373