[deploy] Remove `Package::acquireSession`
Previously, `InterpreterSession` behaved differently depending on who
created it (package vs. interpreter vs. replicatedobj). This is pretty
confusing to people, and was one of the major motivations for removing
`self` in https://github.com/pytorch/pytorch/pull/66377.

Now that `InterpreterSession` only behaves one way, we don't need the
`acquireSession` methods defined on Package or ReplicatedObj. This
PR removes the one for Package.

ghstack-source-id: e4e87b35cdabc8b6264f682f111f9dd0c69256ee
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66533