[deploy] remove ReplicatedObj::acquireSession
Previously, `InterpreterSession` behaved differently depending on who
created it (package vs. interpreter vs. replicatedobj). This is pretty
confusing to people, and was one of the major motivations for removing
`self` in https://github.com/pytorch/pytorch/pull/66377.

Now that `InterpreterSession` only behaves one way, we don't need the
`acquireSession` methods defined on Package or ReplicatedObj. This
  PR removes the one for ReplicatedObj.

Some places try to acquire a session from a ReplicatedObj without having a
reference to the InterpreterManager. This is suspicious (if you don't know the
lifetime of the InterpreterManager, bad things could happen if you arbitrarily
try to schedule work on it). But I expose a ReplicatedObj::getManager to compensate.

ghstack-source-id: 27bbfc7ae10d607adbc4ba08ab17ae35ac7e9e5b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66534