[2/n] Put native message types into BuiltinMessageType enum
Context: to support custom messages, the `MessageType` field in RPC
Message can no longer be an enum. We plan to change that to an
`unint16_t` and allow users to register their own types.

This commit makes `MessageType` as a alias of `uint16_t`, and
rename existing `MessageType` to `BuiltinMessageType`.

ghstack-source-id: 39bb65c8c30f6ccf447bd13b90da196d1d9e81f3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68089