dbr quant: fix debugging fqn info for converted model
Summary:

Fixes the debugging FQN info for a converted model. Some of this
information was missing because eager mode convert performed
module swaps. This information is only used in debugging and is
not used for inference.

Test plan:

```
python test/test_quantization.py TestQuantizeDBR
```

turn `enable_logging` on in `auto_trace.py`, the FQN is now displayed
for a converted model

ghstack-source-id: 693df5ab54b6d8ff788f887ba9623cc54f7fa1ed
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68840