Codegen: Add root_name property to NativeFunction{,sGroup}
This adds `NativeFunction.root_name` which is the canonical name
for the operator group. i.e. the BaseOperatorName without inplace or
double-underscores. In the previous PR I referred to this as
`base_name` but confusingly `BaseOperatorName` does potentially
include inplace or double-underscores.

I also add the property to `NativeFunctionsGroup` so that grouped
functions with type `Union[NativeFunction, NativeFunctionsGroup]`
can have the property queried without needing `isinstance` checks.

ghstack-source-id: 738845f08a3f26214d0f5e6a14df9f3f29a8f70c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68687