Add linalg.lu_factor
This PR exposes `torch.lu` as `torch.linalg.lu_factor` and
`torch.linalg.lu_factor_ex`.

This PR also adds support for matrices with zero elements both in
the size of the matrix and the batch. Note that this function simply
returns empty tensors of the correct size in this case.

We add a test and an OpInfo for the new function.

This PR also adds documentation for this new function in line of
the documentation in the rest of `torch.linalg`.

ghstack-source-id: 6b1195713210ca403f92a90edbee3dd9516ad553
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66933