[quant][graphmode][fx] Add default_replay_qconfig for ops like reshape
Summary:
This PR added default_replay_qconfig and default_replay_observer which is used
when we want to configure an operator to reuse the observer from input, if the input
Tensor for the operator is not observed, we will not observe the output of this operator either,
if the input Tensor is observed, we will observe the output of the operator with the same observer.

e.g.

```
x1 = x0.reshape()
```
if reshape is configured with default_replay_qconfig:
1. if x0 is observed with observer_0, we'll observe x1 with the same observer instance
2. if x0 is not observed, we won't observe x1 either

Test Plan:
```
python test/test_quantization.py TestQuantizeFx.test_replay_qconfig
```
Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 37769ccf6b93a056d2d2bc4a18cee0bfb19b101c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69249