[deploy] docs
This adds some actual documentation for deploy, which is probably useful
since we told everyone it was experimentally available so they will
probably be looking at what the heck it is.

It also wires up various compoenents of the OSS build to actually work
when used from an external project.

ghstack-source-id: 6a5c95676bc2f9f57fba71008dfcc12639769af5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69251